Updated props passing from home.js to project.js to receive project information, moved backend folders to backend.
