--- client/src/components/ProjectContext.js
@@ -1,5 +1,0 @@
-import React from "react";
-
-const ProjectContext = React.createContext({});
-
-export default ProjectContext;
