+++ routes/index.js
@@ -7,7 +7,8 @@ router.use("/api", apiRoutes);
 
 // If no API routes are hit, send the React app
 router.use(function (req, res) {
-    res.sendFile(path.join(__dirname, "../client/build/index.html"));
+    res.sendFile(path.join(__dirname, "../client/index.html"));
+    // res.sendFile(path.join(__dirname, "../client/build/index.html"));
 });
 
 module.exports = router;
