+++ server.js
@@ -2,7 +2,7 @@ const express = require("express");
 const mongoose = require("mongoose");
 
 const app = express();
-const routes = require("./client/src/routes/index.js");
+const routes = require("./routes/index.js");
 const PORT = process.env.PORT || 3001;
 
 // Middleware
